frontend/requirements.txt
# Config Files
python-dotenv

# Frontend
flask
flask-login
flask-wtf
requests